Styled the slider view as well as moved the slider into its own folder in components
