+++ client/src/components/LaunchSlider/LaunchSlider.css
@@ -1,0 +1,73 @@
+.slider{
+    background-image: linear-gradient(navy, black);
+    color: white;
+    width: 100%;
+    margin-bottom: 10%;
+}
+
+.slider-btn-r, .slider-btn-l{
+    border: none;
+    padding: 2%;
+    margin: 5%;
+    background: none;
+    font-weight: bold;
+}
+
+.slider-btn-r{
+    float: right;
+}
+
+.slider-btn-l{
+    float: left;
+}
+
+.slider-btn-r:hover, .slider-btn-l:hover{
+    background: navy;
+}
+
+#rocket-pic{
+    opacity: .75;
+    margin: 5%;
+}
+
+.details-btn{
+    float: right;
+}
+
+.upcoming{
+    margin-left: 35%;
+    margin-top: 5%;
+    color: rgb(177, 177, 177);
+}
+
+.date{
+    font-size: 20px;
+}
+
+.time{
+    font-weight: bold;
+}
+
+.count-down{
+    margin-top: 3%;
+    font-size: 30px;
+    font-weight: lighter;
+    color: rgb(255, 233, 111);
+}
+
+.launch-data{
+    margin: 5%;
+    list-style: none;
+}
+
+.company{
+    font-weight: bold;
+    margin: 1%;
+    margin-left: 0;
+}
+
+.rocket{
+    font-family: sans-serif;
+    font-size: 25px;
+    font-weight: lighter;
+}
