+++ client/src/components/LaunchSlider/LaunchSlider.js
@@ -2,36 +2,39 @@ import React from 'react';
 import Container from 'react-bootstrap/Container';
 import Col from 'react-bootstrap/Col';
 import Row from 'react-bootstrap/Row';
-import FollowButton from '../components/button';
+import FollowButton from '../button';
 import Button from 'react-bootstrap/Button';
-import Counter from '../components/Counter';
-import DetailButton from '../components/DetailButton';
-import Weather from '../components/Weather';
+import Counter from '../Counter';
+import DetailButton from '../DetailButton';
+import Weather from '../Weather';
+import moment from 'moment';
+import "./LaunchSlider.css";
 
 const LaunchSlider = (props) => {
     const {launch, prevDate, nextDate, total, handleIndexChange, index} = props
     // console.log("LAUNCH SLIDER", launch.countdownTime)
     // console.log(launch.company)
     return(
-        <Container className="slider" style={{backgroundColor: "blue"}}>
+        <Container className="slider">
             <Row>
                 <Col>
                     <Container>
                         <Row>
                             <Col>
-                                <ul>
-                                    <li>Location: {launch.location}</li>
-                                    <li>Date: {launch.date} </li>
+                                <ul className="launch-data">
+                                    <li className="date">{moment(launch.date).format("MMMM DD, YYYY")} </li>
+                                    <li className="time">{moment(launch.date).format("LTS")} </li>
                                     <Weather location={launch.location} />
-                                    <Counter date={`${launch.countdownTime}`} />
+                                    <li className="count-down"><Counter date={`${launch.countdownTime}`} /></li>
+                                    <li className="location">{launch.location}</li>
                                 </ul>
                             </Col>
                         </Row>
                         <Row>
                             <Col>
-                               <ul>
-                                   <li>Company: {launch.company}</li>
-                                   <li>Rocket: {launch.rocket}</li>
+                               <ul className="launch-data">
+                                   <li className="company"> {launch.company}</li>
+                                   <li className="rocket">{launch.rocket}</li>
                                </ul>
                             </Col>
                         </Row>
@@ -43,7 +46,7 @@ const LaunchSlider = (props) => {
             </Row>
             <Row>
                 <Col>
-                    <DetailButton index={index} />
+                    <DetailButton className="details-btn" index={index} />
                 </Col>
                 <Col>
                     <FollowButton text="Follow" />
@@ -51,13 +54,14 @@ const LaunchSlider = (props) => {
             </Row>
             <Row>
                 <Col>
-                    <Button onClick={()=> handleIndexChange(-1)}> {prevDate} </Button>
+                    
+                {prevDate !== 'none' ? <Button className="slider-btn-l" onClick={()=> handleIndexChange(-1)}> {prevDate} </Button> : ''}
                 </Col>
                 <Col>
-                    <h4>{total} Upcoming</h4>
+                    <p className="upcoming">{total} Launches</p>
                 </Col>
                 <Col>
-                    <Button onClick={()=> handleIndexChange(1)}> {nextDate} </Button>
+                {nextDate !== 'none' ? <Button className="slider-btn-r" onClick={()=> handleIndexChange(1)}> {nextDate} </Button> : ''}
                 </Col>
             </Row>
         </Container>
