+++ client/src/components/button/style.css
@@ -12,4 +12,9 @@
     border-bottom-right-radius: 50px;
     border-top-right-radius: 50px;
     border-left: .5px solid white;
+    float: left;
+}
+
+.follow:hover{
+    color: rgb(160, 160, 160);
 }
