+++ client/src/pages/LaunchDetails.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import Template from '../components/template/template.wrapper';
-import LaunchSlider from '../components/LaunchSlider';
+import LaunchSlider from '../components/LaunchSlider/LaunchSlider';
 import ArticleList from '../components/ArticleList';
 import { connect } from 'react-redux';
